Fix lint build and refactor routes
