+++ pattern-library/docs/src/components/Routes/generateRoutes.jsx
