+++ pattern-library/docs/src/doc-components/PropTypesTable/utils.jsx
@@ -1,5 +1,5 @@
 import React, { PropTypes, createElement } from 'react';
-import { parseType } from '#docs/doc-components/utilities';
+import { parseType } from '#docs/utils/prop-types';
 
 const TypeShape = PropTypes.shape({
   name: PropTypes.string,
