+++ pattern-library/docs/src/index.html
@@ -5,6 +5,6 @@
     <title>Webpack App</title>
   </head>
   <body>
-    <div id="app"></div>
+    <div id="root"></div>
   </body>
 </html>
