+++ pattern-library/docs/src/utils/prop-types.js
@@ -1,5 +1,3 @@
-/* eslint-disable import/prefer-default-export */
-
 export function parseType({ name, raw, value } = {}) {
   if (name === 'shape') {
     const nestedProps = Object.keys(value).map(prop =>
